Remove unnecessary u32 cast
Having switched the seed from u64 to u32, this is no longer necessary.

+++ src/world.rs
@@ -53,7 +53,7 @@ impl Chunk {
         // NOTE: `x` and `y` overflow on negative values here, but all
         // we care about is having a distinct value for each position
         // so our seeds don't repeat. So this is fine here.
-        let seed = Wrapping(u32::from(world_seed))
+        let seed = Wrapping(world_seed)
             + (Wrapping(13) * Wrapping(pos.x as u32))
             + (Wrapping(17) * Wrapping(pos.y as u32));
